refactor: 💡 스케쥴 도메인의 스케쥴 자동생성중 정렬정기일정은 정기일정 도메인으로 변경
